move findAndReplace func in its own file
